// src/app/enums/apiResponse.ts
export interface WebAppUserProfileResp {
  statuscode: number;
  msg: string;
  name: string;
  outletName: string;
  emailID: string;
  mobileNo: string;
  alternateMobile: string;
  dob: string;
  pan: string;
  pincode: string;
  city: string;
  state: string;
  address: string;
  aadhar: string;
  profilePic: string;
}